# ProductTraceability_2026/doc/openapi.yaml
openapi: 3.0.3
info:
  title: Agricultural Traceability System API
  version: 1.0.0
  description: API specification for the agricultural product traceability system.
servers:
  - url: '{baseUrl}'
    variables:
      baseUrl:
        default: http://localhost:8888
paths:
...
  /api/trace/detail:
    get:
      summary: "[Full] Trace detail with full lifecycle timeline"
      description: |
        Aggregates data from trace_batch, trace_record, trace_inspection,
        trace_processing, trace_storage, trace_transport, trace_sale.
        Returns batchInfo + timeline[] + enterpriseInfo.
        Each timeline node: { stage, title, time, operator, txHash, blockNumber, details{}, images[], official? }
      tags: [Trace]
      parameters:
        - { name: batchId, in: query, required: true, schema: { type: string }, description: "batch ID or batch code" }
      responses:
        '200':
          description: Aggregated trace timeline
  /api/blockchain/verify:
    post:
      summary: "[Full] Verify blockchain data integrity"
      tags: [Trace]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                batchId: { type: integer }
                batchCode: { type: string }
      responses:
        '200':
          description: Verification result
...
  /api/system/user/list:
    get:
      summary: "✅ User list with pagination"
      tags: [System-User]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: current, in: query, schema: { type: integer, default: 1 } }
        - { name: size, in: query, schema: { type: integer, default: 10 } }
        - { name: username, in: query, schema: { type: string } }
        - { name: phone, in: query, schema: { type: string } }
        - { name: keyword, in: query, schema: { type: string }, description: "Search by username/realName/phone" }
        - { name: userType, in: query, schema: { type: integer } }
        - { name: status, in: query, schema: { type: integer } }
      responses:
        '200':
          description: Paginated user list (password excluded)
  /api/system/user/add:
    post:
      summary: "✅ Add user"
      tags: [System-User]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username]
              properties:
                username: { type: string }
                realName: { type: string }
                password: { type: string, description: "Defaults to 123456 if empty" }
                phone: { type: string }
                email: { type: string }
                userType: { type: integer }
                enterpriseId: { type: integer }
                roleIds: { type: array, items: { type: integer } }
                remark: { type: string }
      responses:
        '200':
          description: Success
  /api/system/user/update:
    put:
      summary: "✅ Update user"
      tags: [System-User]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id]
              properties:
                id: { type: integer }
                realName: { type: string }
                phone: { type: string }
                email: { type: string }
                userType: { type: integer }
                enterpriseId: { type: integer }
                roleIds: { type: array, items: { type: integer } }
                remark: { type: string }
      responses:
        '200':
          description: Success
  /api/system/user/delete/{id}:
    delete:
      summary: "✅ Delete user"
      tags: [System-User]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Success
  /api/system/user/reset-password/{id}:
    put:
      summary: "✅ Reset user password to 123456"
      tags: [System-User]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Password reset
  /api/system/user/toggle-status/{id}:
    put:
      summary: "✅ Toggle user status"
      tags: [System-User]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: status, in: query, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Success

  # ==========================================
  # System Role Management
  # ==========================================
...
  /api/system/user/change-password:
    put:
      summary: "✅ User change password (verify old password)"
      tags: [System-User]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [oldPassword, newPassword]
              properties:
                oldPassword: { type: string }
                newPassword: { type: string }
      responses:
        '200':
          description: Password changed

  # ==========================================
  # Message Notification
  # ==========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...